LessonInformation 모델 및 누락된 프로젝트 파일 추가
