+++ Nanagong/Nanagong/RegisterLessonView/Model/RegisterLessionViewModel.swift
@@ -37,23 +37,6 @@ enum RegisterLessionViewNavigationType {
     case nextStep
 }
 
-struct LessonInformation {
-    
-    static let empty: Self = .init(lessonName: "", totalNumber: 0, categoryId: 0, siteId: 0, alertDays: nil, startDate: Date(), endDate: Date(), message: nil, price: 0)
-    
-    var lessonName: String
-    var totalNumber: Int
-    var categoryId: Int
-    var categoryName: String?
-    var siteId: Int
-    var siteName: String?
-    var alertDays: String?
-    var startDate: Date
-    var endDate: Date
-    var message: String?
-    var price: Int
-}
-
 final class RegisterLessionViewModel {
     
     @Published var nextButtonState: ButtonState
